+++ template_sets.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"os"
 	"sync"
+	"sync/atomic"
 )
 
 // TemplateLoader allows to implement a virtual file system.
@@ -47,7 +48,7 @@ type TemplateSet struct {
 	// For efficiency reasons you can ban tags/filters only *before* you have
 	// added your first template to the set (restrictions are statically checked).
 	// After you added one, it's not possible anymore (for your personal security).
-	firstTemplateCreated bool
+	firstTemplateCreated atomic.Bool
 	bannedTags           map[string]bool
 	bannedFilters        map[string]bool
 
@@ -101,7 +102,7 @@ func (set *TemplateSet) BanTag(name string) error {
 	if !has {
 		return fmt.Errorf("tag '%s' not found", name)
 	}
-	if set.firstTemplateCreated {
+	if set.firstTemplateCreated.Load() {
 		return errors.New("you cannot ban any tags after you've added your first template to your template set")
 	}
 	_, has = set.bannedTags[name]
@@ -125,7 +126,7 @@ func (set *TemplateSet) BanFilter(name string) error {
 	if !has {
 		return fmt.Errorf("filter '%s' not found", name)
 	}
-	if set.firstTemplateCreated {
+	if set.firstTemplateCreated.Load() {
 		return errors.New("you cannot ban any filters after you've added your first template to your template set")
 	}
 	_, has = set.bannedFilters[name]
@@ -149,7 +150,7 @@ func (set *TemplateSet) resolveTemplate(tpl *Template, path string) (name string
 		name = set.resolveFilenameForLoader(loader, tpl, path)
 		fd, err = loader.Get(name)
 		if err == nil {
-			return
+			return name, loader, fd, err
 		}
 	}
 
@@ -206,21 +207,21 @@ func (set *TemplateSet) FromCache(filename string) (*Template, error) {
 
 // FromString loads a template from string and returns a Template instance.
 func (set *TemplateSet) FromString(tpl string) (*Template, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	return newTemplateString(set, []byte(tpl))
 }
 
 // FromBytes loads a template from bytes and returns a Template instance.
 func (set *TemplateSet) FromBytes(tpl []byte) (*Template, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	return newTemplateString(set, tpl)
 }
 
 // FromFile loads a template from a filename and returns a Template instance.
 func (set *TemplateSet) FromFile(filename string) (*Template, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	_, _, fd, err := set.resolveTemplate(nil, filename)
 	if err != nil {
@@ -244,7 +245,7 @@ func (set *TemplateSet) FromFile(filename string) (*Template, error) {
 
 // RenderTemplateString is a shortcut and renders a template string directly.
 func (set *TemplateSet) RenderTemplateString(s string, ctx Context) (string, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	tpl := Must(set.FromString(s))
 	result, err := tpl.Execute(ctx)
@@ -256,7 +257,7 @@ func (set *TemplateSet) RenderTemplateString(s string, ctx Context) (string, err
 
 // RenderTemplateBytes is a shortcut and renders template bytes directly.
 func (set *TemplateSet) RenderTemplateBytes(b []byte, ctx Context) (string, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	tpl := Must(set.FromBytes(b))
 	result, err := tpl.Execute(ctx)
@@ -268,7 +269,7 @@ func (set *TemplateSet) RenderTemplateBytes(b []byte, ctx Context) (string, erro
 
 // RenderTemplateFile is a shortcut and renders a template file directly.
 func (set *TemplateSet) RenderTemplateFile(fn string, ctx Context) (string, error) {
-	set.firstTemplateCreated = true
+	set.firstTemplateCreated.Store(true)
 
 	tpl := Must(set.FromFile(fn))
 	result, err := tpl.Execute(ctx)